fix: fix bug empty searchTerm
