+++ app/search/[id].jsx
@@ -1,20 +1,14 @@
 import { Stack, useRouter, useSearchParams } from 'expo-router';
-import { ActivityIndicator, RefreshControl, SafeAreaView, ScrollView, Text, View } from 'react-native';
-import useFetch from '../../hook/useFetch';
-import useMyFetch from '../../hook/useMyFetch';
-import useJsonOneJob from '../../hook/useJsonOneJob';
-import { COLORS, SIZES, icons } from '../../constants';
-import { Company, JobFooter, JobTabs, ScreenHeaderBtn, Specifics } from '../../components';
-import { useEffect, useState } from 'react';
-import JobSearch from '../../components/home/search/JobSearch';
+import { SafeAreaView } from 'react-native';
+import { COLORS, icons } from '../../constants';
+import { ScreenHeaderBtn, JobSearch } from '../../components';
+import { useState } from 'react';
 import React from 'react';
 
 
 const SearchResults = () => {
   const params = useSearchParams();
   const router = useRouter();
-  const [refreshing, setRefreshing] = useState(false);
-  const onRefresh = () => { };
 
   return (
     <SafeAreaView
