+++ components/home/welcome/Welcome.jsx
@@ -6,7 +6,7 @@ import {
   Image,
   FlatList
 } from 'react-native';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import styles from './welcome.style';
 import { icons, SIZES } from '../../../constants';
 import { useRouter } from 'expo-router';
@@ -36,7 +36,7 @@ const Welcome = ({ searchTerm, setSearchTerm, handleClick }) => {
             placeholder='What are you looking for'
           />
         </View>
-        <TouchableOpacity style={styles.searchBtn} onPress={() => { handleClick(); }}>
+        <TouchableOpacity style={styles.searchBtn} onPress={handleClick}>
           <Image
             source={icons.search}
             resizeMode='contain'
